dom/tests/mochitest/general/test_bug1434273.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1434273
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1434273</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
    #test::before { content: url(image_200.png); }
  </style>
  <script type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  /** Test for Bug 1434273 **/
  addLoadEvent(function() {
    synthesizeMouse($("test"), 100, 100, {});
    ok(true, "We did not crash!");
    SimpleTest.finish();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1434273">Mozilla Bug 1434273</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
